app/requirements.txt
playwright
rich
telnyx
python-dotenv
yarl
flask
cssselect
lxml
mysql-connector-python
pymysql
cryptography
tortoise-orm
aiomysql